python_metadata/requirements.txt
eyed3
requests
pillow
tqdm
mutagen
pydub
colorthief